+++ betterrandom/src/main/java/io/github/pr0methean/betterrandom/prng/adapter/ReseedingSplittableRandomAdapter.java
@@ -8,8 +8,10 @@ import io.github.pr0methean.betterrandom.seed.SimpleRandomSeeder;
 import io.github.pr0methean.betterrandom.util.Java8Constants;
 import java.util.Collections;
 import java.util.Map;
+import java.util.Objects;
 import java.util.WeakHashMap;
 import java8.util.SplittableRandom;
+import javax.annotation.Nullable;
 
 /**
  * Like {@link SplittableRandomAdapter}, but uses a {@link SimpleRandomSeeder} to replace each
@@ -40,7 +42,7 @@ public class ReseedingSplittableRandomAdapter extends BaseSplittableRandomAdapte
    * @throws SeedException if {@code seedGenerator} fails to generate an initial seed
    */
   protected ReseedingSplittableRandomAdapter(final SeedGenerator seedGenerator,
-      SimpleRandomSeeder randomSeeder) throws SeedException {
+      @Nullable SimpleRandomSeeder randomSeeder) throws SeedException {
     super(seedGenerator.generateSeed(Java8Constants.LONG_BYTES));
     this.seedGenerator = seedGenerator;
     this.randomSeeder.set(randomSeeder);
@@ -74,7 +76,8 @@ public class ReseedingSplittableRandomAdapter extends BaseSplittableRandomAdapte
    *     seed.
    */
   @SuppressWarnings("SynchronizationOnStaticField")
-  public static ReseedingSplittableRandomAdapter getInstance(final SimpleRandomSeeder randomSeeder,
+  public static ReseedingSplittableRandomAdapter getInstance(
+      @Nullable final SimpleRandomSeeder randomSeeder,
       final SeedGenerator seedGenerator) throws SeedException {
     ReseedingSplittableRandomAdapter instance = INSTANCES.get(randomSeeder);
     if (instance == null) {
@@ -97,8 +100,8 @@ public class ReseedingSplittableRandomAdapter extends BaseSplittableRandomAdapte
     return threadLocal.get().getSeed();
   }
 
-  @Override public void setRandomSeeder(final SimpleRandomSeeder randomSeeder) {
-    if (!this.randomSeeder.get().equals(randomSeeder)) {
+  @Override public void setRandomSeeder(@Nullable final SimpleRandomSeeder randomSeeder) {
+    if (!Objects.equals(this.randomSeeder.get(), randomSeeder)) {
       throw new UnsupportedOperationException(
           "ReseedingSplittableRandomAdapter's binding to LegacyRandomSeeder is immutable");
     }
@@ -155,10 +158,10 @@ public class ReseedingSplittableRandomAdapter extends BaseSplittableRandomAdapte
       return false;
     }
     ReseedingSplittableRandomAdapter that = (ReseedingSplittableRandomAdapter) o;
-    return randomSeeder.get().equals(that.randomSeeder.get());
+    return Objects.equals(randomSeeder.get(), that.randomSeeder.get());
   }
 
   @Override public int hashCode() {
-    return randomSeeder.hashCode() + 1;
+    return Objects.hash(randomSeeder.get());
   }
 }