+++ betterrandom/src/main/java/io/github/pr0methean/betterrandom/seed/SimpleRandomSeeder.java
@@ -11,7 +11,6 @@ import java.io.Serializable;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Random;
 import java.util.Set;
@@ -27,8 +26,7 @@ import java.util.concurrent.locks.Condition;
  * @author Chris Hennick
  */
 public class SimpleRandomSeeder extends Looper {
-  private static final Map<ByteArrayReseedableRandom, byte[]> SEED_ARRAYS =
-      Collections.synchronizedMap(new WeakHashMap<ByteArrayReseedableRandom, byte[]>(1));
+
   // FIXME: Setting a longer POLL_INTERVAL slows many tests, and causes some to time out
   // (Why doesn't BaseRandom's call to wakeUp() prevent this?!)
   /**
@@ -46,8 +44,6 @@ public class SimpleRandomSeeder extends Looper {
    * The seed generator this seeder uses.
    */
   protected final SeedGenerator seedGenerator;
-
-  private final byte[] longSeedArray = new byte[8];
 
   /**
    * Holds {@link ByteArrayReseedableRandom} instances that should be reseeded when their entropy is
@@ -263,14 +259,7 @@ public class SimpleRandomSeeder extends Looper {
         if (random.preferSeedWithLong()) {
           reseedWithLong((Random) random);
         } else {
-          byte[] seedArray1 = SEED_ARRAYS.get(random);
-          if (seedArray1 == null) {
-            seedArray1 = new byte[random.getNewSeedLength()];
-            SEED_ARRAYS.put(random, seedArray1);
-          }
-          byte[] seedArray = seedArray1;
-          seedGenerator.generateSeed(seedArray);
-          random.setSeed(seedArray);
+          random.setSeed(seedGenerator.generateSeed(random.getNewSeedLength()));
         }
       }
     } finally {
@@ -284,8 +273,7 @@ public class SimpleRandomSeeder extends Looper {
    * @param random the PRNG to reseed
    */
   protected void reseedWithLong(final Random random) {
-    seedGenerator.generateSeed(longSeedArray);
-    random.setSeed(BinaryUtils.convertBytesToLong(longSeedArray));
+    random.setSeed(BinaryUtils.convertBytesToLong(seedGenerator.generateSeed(Long.BYTES)));
   }
 
   /**