+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.3.0]
+
 ### Changed
 - `RandomDotOrgSeedGenerator` now uses Commons-Codec to decode Base 64, which should fix a possible crash on Android.
 - `SplittableRandomAdapter` is now deprecated. It can be replaced by either `ThreadLocalRandom` or
@@ -573,8 +575,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release on Maven Central.
 
-[Unreleased]: https://github.com/Pr0methean/BetterRandom/compare/BetterRandom-Java7-4.2.1...master
-[4.2.1]: https://github.com/Pr0methean/BetterRandom/compare/BetterRandom-Java7-4.2.0...BetterRandom-Java7-4.2.1
+[Unreleased]: https://github.com/Pr0methean/BetterRandom/compare/BetterRandom-Java7-4.3.0...master
+[4.3.0]: https://github.com/Pr0methean/BetterRandom/compare/BetterRandom-Java7-4.2.0...BetterRandom-Java7-4.3.0
 [4.2.0]: https://github.com/Pr0methean/BetterRandom/compare/BetterRandom-Java7-4.1.0...BetterRandom-Java7-4.2.0
 [4.1.0]: https://github.com/Pr0methean/BetterRandom/compare/BetterRandom-Java7-4.0.3...BetterRandom-Java7-4.1.0
 [4.0.3]: https://github.com/Pr0methean/BetterRandom/compare/BetterRandom-Java7-4.0.2...BetterRandom-Java7-4.0.3